# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bank Statement Analyzer Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # Statements
...
  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get dashboard summary - total income, expenses, savings, health score
      parameters:
        - name: month
          in: query
          schema:
            type: ["string", "null"]
          description: "YYYY-MM format. Defaults to current month"
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
  /analytics/spending-by-category:
    get:
      operationId: getSpendingByCategory
      tags: [analytics]
      summary: Spending breakdown by category
      parameters:
        - name: month
          in: query
          schema:
            type: ["string", "null"]
        - name: period
          in: query
          schema:
            type: ["string", "null"]
            enum: [month, quarter, year, null]
      responses:
        "200":
          description: Category spending breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategorySpending"
  /analytics/cashflow:
    get:
      operationId: getCashflow
      tags: [analytics]
      summary: Monthly income vs expenses trend
      parameters:
        - name: months
          in: query
          schema:
            type: ["integer", "null"]
          description: Number of months (default 6)
      responses:
        "200":
          description: Cashflow trend
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CashflowMonth"
  /analytics/merchant-insights:
    get:
      operationId: getMerchantInsights
      tags: [analytics]
      summary: Top merchants by spend
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Merchant insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MerchantInsight"
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserPage"
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/audit-logs:
    get:
      operationId: listAuditLogs
      tags: [admin]
      summary: List audit logs
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
        name:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role, plan]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin, super_admin]
        plan:
          type: string
          enum: [free, pro, enterprise]
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
...
    AnalyticsSummary:
      type: object
      required: [totalIncome, totalExpenses, netSavings, savingsRate, healthScore, transactionCount, month]
      properties:
        totalIncome:
          type: number
        totalExpenses:
          type: number
        netSavings:
          type: number
        savingsRate:
          type: number
        healthScore:
          type: number
        transactionCount:
          type: integer
        month:
          type: string
        topCategory:
          type: ["string", "null"]
        budgetUtilization:
          type: ["number", "null"]
    CategorySpending:
      type: object
      required: [categoryId, categoryName, color, icon, amount, percentage, count]
      properties:
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: string
        color:
          type: string
        icon:
          type: string
        amount:
          type: number
        percentage:
          type: number
        count:
          type: integer
    CashflowMonth:
      type: object
      required: [month, income, expenses, net]
      properties:
        month:
          type: string
        income:
          type: number
        expenses:
          type: number
        net:
          type: number
    MerchantInsight:
      type: object
      required: [merchant, totalSpent, transactionCount, averageSpend]
      properties:
        merchant:
          type: string
        totalSpent:
          type: number
        transactionCount:
          type: integer
        averageSpend:
          type: number
        lastTransaction:
          type: ["string", "null"]
...
    AdminUserPage:
      type: object
      required: [users, total]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: ["string", "null"]
          enum: [user, admin, super_admin, null]
        plan:
          type: ["string", "null"]
          enum: [free, pro, enterprise, null]
        status:
          type: ["string", "null"]
          enum: [active, suspended, null]
    AuditLog:
      type: object
      required: [id, userId, action, entity, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        userEmail:
          type: ["string", "null"]
        action:
          type: string
        entity:
          type: string
        entityId:
          type: ["integer", "null"]
        details:
          type: ["string", "null"]
        createdAt:
          type: string