# lib/api-zod/src/generated/types/cashflowMonth.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bank Statement Analyzer Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface CashflowMonth {
  month: string;
  income: number;
  expenses: number;
  net: number;
}